Conflicts:
	src/byui/cit260/roadToSaltLake/model/Game.java
	src/byui/cit260/roadToSaltLake/model/Resources.java
	src/byui/cit260/roadToSaltLake/view/OccupationView.java
	src/byui/cit260/roadToSaltLake/view/StoreView.java

+++ src/byui/cit260/roadToSaltLake/control/Store.java
@@ -11,9 +11,9 @@ package byui.cit260.roadToSaltLake.control;
  */
 public class Store {    
     
-    public float purchaseInventory(float funds, float itemQuantity, float itemPrice)
+    public double purchaseInventory(double funds, double itemQuantity, double itemPrice)
     {   
-        float totalCost;
+        double totalCost;
         
         if (funds <= 0)
         {
